+++ Source/Core/Common/Common.vcxproj.filters
@@ -25,7 +25,6 @@
     <ClInclude Include="BitField.h" />
     <ClInclude Include="BitSet.h" />
     <ClInclude Include="BlockingLoop.h" />
-    <ClInclude Include="BreakPoints.h" />
     <ClInclude Include="CDUtils.h" />
     <ClInclude Include="ChunkFile.h" />
     <ClInclude Include="CodeBlock.h" />
@@ -229,7 +228,6 @@
     <ClInclude Include="Semaphore.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="BreakPoints.cpp" />
     <ClCompile Include="CDUtils.cpp" />
     <ClCompile Include="ColorUtil.cpp" />
     <ClCompile Include="ENetUtil.cpp" />
